Reduce extend of FileActivity, use DrawerActivity instead
